move ArrayColumn into Column
